# 🚀 MAJOR SEV-0: Upgrade Next.js to 15.5.0