--- a/CLAUDE.md
+++ b/CLAUDE.md
@@ -2,11 +2,13 @@
 **🎖️ BRTOPS ENABLED** - Military-precision development operations active
 
 ## BRTOPS Integration Status
-✅ **Framework Version**: 1.0.000  
+✅ **Framework Version**: 1.1.0 - Structure & Workflow Enhancement  
 ✅ **Command Structure**: GO RCC, GO PLAN, GO CODE, GO FINAL, GO VAL, DEBRIEF  
+✅ **Enhanced Commands**: INIT FEATURE, CREATE BRANCH, MERGE READY, PREFLIGHT, NOFLIGHT  
+✅ **Project Management**: 6-folder documentation structure, git workflow integration  
 ✅ **Guide Mode**: GUIDE AUTO (adaptive to user experience)  
 ✅ **Collaboration Mode**: COLLAB (equal partnership default)  
-✅ **Quality Gates**: Enabled for all SEV levels  
+✅ **Quality Gates**: Enabled for all SEV levels with automated enforcement  
 
 ## Important Project Information