--- a/CLAUDE.md
+++ b/CLAUDE.md
@@ -2,13 +2,14 @@
 **🎖️ BRTOPS ENABLED** - Military-precision development operations active
 
 ## BRTOPS Integration Status
-✅ **Framework Version**: 1.1.0 - Structure & Workflow Enhancement  
+✅ **Framework Version**: 1.1.003 - Enhanced 6-Folder Structure Integration  
 ✅ **Command Structure**: GO RCC, GO PLAN, GO CODE, GO FINAL, GO VAL, DEBRIEF  
-✅ **Enhanced Commands**: INIT FEATURE, CREATE BRANCH, MERGE READY, PREFLIGHT, NOFLIGHT  
-✅ **Project Management**: 6-folder documentation structure, git workflow integration  
+✅ **Enhanced Commands**: INIT FEATURE, CREATE DOCS, VALIDATE DOCS, GENERATE TEMPLATES  
+✅ **Documentation System**: Enhanced 6-folder structure with SEV-level requirements  
+✅ **Project Management**: Complete template library, git workflow integration  
 ✅ **Guide Mode**: GUIDE AUTO (adaptive to user experience)  
 ✅ **Collaboration Mode**: COLLAB (equal partnership default)  
-✅ **Quality Gates**: Enabled for all SEV levels with automated enforcement  
+✅ **Quality Gates**: Enhanced documentation completeness validation by SEV level  
 
 ## Important Project Information
 
@@ -195,4 +196,26 @@ npm run dev
 
 *Last updated: 2025-08-21*
 *This file helps Claude Code remember important project details across sessions*
-*🎖️ BRTOPS Framework v1.0.000 integrated - Military-precision development operations active*
+*🎖️ BRTOPS Framework v1.1.003 integrated - Enhanced 6-folder structure with military-precision development operations*
+
+## Enhanced 6-Folder Documentation Structure (v1.1.003)
+
+### Project Excellence Standards
+This project exemplifies BRTOPS enhanced documentation standards with:
+- **Enhanced Structure**: `01-objectives/`, `02-analysis/`, `03-architecture-design/`, `04-development/`, `05-debugging/`, `06-key_learnings/`
+- **100% BRTOPS Compliance**: Complete infrastructure with `.brtops/` configuration
+- **Quality Gates**: SEV-specific documentation requirements implemented
+- **Template Integration**: Full template library for consistent documentation
+
+### Documentation Commands Available
+```
+CREATE DOCS SEV-[X] [feature-name]   - Generate enhanced 6-folder structure
+VALIDATE DOCS [feature-name]          - Check documentation completeness
+CHECK DOCS SEV-[X]                   - Validate SEV-level compliance
+GENERATE TEMPLATES [folder]           - Create required templates
+```
+
+### BRTOPS Configuration Location
+- **Global Config**: `/Users/bthompso/.claude/CLAUDE.md` (v1.1.003)
+- **Project Config**: `.brtops/config.json` (project-specific settings)
+- **Feature Registry**: `.brtops/feature-registry.json` (feature tracking)