--- a/data/coverage-data.js
+++ b/data/coverage-data.js
@@ -4,7 +4,7 @@
  */
 
 export const COVERAGE_DATA = {
-                                                                                                                          summary: {
+                                                                                                                            summary: {
     statements: { covered: 170, total: 200, percentage: 85.20 },
     branches: { covered: 107, total: 120, percentage: 89.50 },
     functions: { covered: 55, total: 70, percentage: 78.80 },
@@ -204,6 +204,7 @@ export const COVERAGE_DATA = {
   ],
   
   testHistory: [
+    { date: '2025-08-21T19:54:16.470Z', totalTests: 623, passed: 460, failed: 163, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
     { date: '2025-08-21T18:30:20.534Z', totalTests: 623, passed: 463, failed: 160, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
     { date: '2025-08-21T16:12:21.475Z', totalTests: 623, passed: 463, failed: 160, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
     { date: '2025-08-21T15:41:16.830Z', totalTests: 604, passed: 442, failed: 162, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
@@ -233,7 +234,7 @@ export const COVERAGE_DATA = {
     { date: '2025-08-20T19:43:24.610Z', totalTests: 571, passed: 409, failed: 162, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
     { date: '2025-08-20T19:32:50.072Z', totalTests: 571, passed: 395, failed: 176, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
     { date: '2025-08-20T18:42:34.666Z', totalTests: 571, passed: 408, failed: 163, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
-    { date: '2025-08-20T18:33:20.203Z', totalTests: 571, passed: 408, failed: 163, duration: 0, coverage: { statements: 85.2, branches: 89.5, functions: 78.8, lines: 85.2 } },
+