--- a/docs/02_features/feature-index.md
+++ b/docs/02_features/feature-index.md
@@ -1,0 +1,104 @@
+# Feature Index - Conversation Tracker
+
+## Overview
+This document provides a master registry of all features implemented in the conversation-tracker project, following BRTOPS v1.1.002 documentation standards.
+
+## Feature Summary
+- **Total Features**: 9 completed
+- **Active Development**: 0 features  
+- **Completed Features**: 9 features
+- **Features with AARs**: 3 features
+
+## Feature Registry
+
+### Major Features (SEV-0/1)
+
+#### SSE Real-Time Collaboration **[SEV-0]**
+- **Path**: `sse-real-time-collaboration/`
+- **Type**: MAJOR FEATURE ENHANCEMENT  
+- **Status**: ✅ Completed (2025-08-21)
+- **Description**: Comprehensive real-time collaboration system with multi-tab synchronization
+- **AAR**: `../04_after-action-reports/sse-multi-tab-synchronization_2025-08-21.md`
+
+#### Guest System Expansion **[SEV-1]**
+- **Path**: `guest-system-expansion/`
+- **Type**: MAJOR FEATURE ENHANCEMENT
+- **Status**: ✅ Completed (2025-08-19)
+- **Description**: Expanded guest user system with enhanced avatar management
+
+#### React Query Migration **[SEV-1]**
+- **Path**: `react-query-migration/`
+- **Type**: MAJOR SYSTEM REFACTOR
+- **Status**: ✅ Completed (2025-08-18)
+- **Description**: Migration from legacy hooks to React Query for improved data management
+
+### Minor Features (SEV-2/3)
+
+#### Card Flip SSE Safe Implementation **[SEV-2]**
+- **Path**: `card-flip-sse-safe-implementation/`
+- **Type**: MINOR FEATURE ENHANCEMENT
+- **Status**: ✅ Completed (2025-08-21)
+- **Description**: Safe card flipping with SSE integration
+
+#### SSE Conversation Polling Elimination **[SEV-2]**
+- **Path**: `sse-conversation-polling-elimination/`
+- **Type**: MINOR OPTIMIZATION
+- **Status**: ✅ Completed (2025-08-19)
+- **Description**: Eliminated polling in favor of SSE-based conversation updates
+
+#### Unified User Management **[SEV-2]**
+- **Path**: `unified-user-management/`
+- **Type**: MINOR SYSTEM ENHANCEMENT
+- **Status**: ✅ Completed (2025-08-18)  
+- **Description**: Consolidated user management across app and dev environments
+
+#### User Selectable Themes **[SEV-3]**
+- **Path**: `user-selectable-themes/`
+- **Type**: MINOR FEATURE ENHANCEMENT
+- **Status**: ✅ Completed (2025-08-18)
+- **Description**: User-configurable color theme selection system
+
+#### User Theme Mode Isolation **[SEV-3]**
+- **Path**: `user-theme-mode-isolation/`
+- **Type**: MINOR FEATURE ENHANCEMENT  
+- **Status**: ✅ Completed (2025-08-21)
+- **Description**: Isolated theme mode management per user
+- **AAR**: `../04_after-action-reports/user-theme-mode-isolation_2025-08-21.md`
+
+### Tweaks (SEV-4/5)
+
+#### Dev Scripts Visual Polish **[SEV-4]**
+- **Path**: `dev-scripts-visual-polish/`
+- **Type**: TWEAK ENHANCEMENT
+- **Status**: ✅ Completed (2025-08-21)
+- **Description**: Visual improvements to development script interfaces
+- **AAR**: `../04_after-action-reports/dev-scripts-visual-polish_2025-08-21.md`
+
+#### User Presence Real-Time **[SEV-4]**
+- **Path**: `user-presence-real-time/`
+- **Type**: TWEAK ENHANCEMENT
+- **Status**: ✅ Completed (2025-08-20)
+- **Description**: Real-time user presence indicators
+
+## Documentation Standards
+
+All features follow the BRTOPS 6-folder structure:
+1. **1-requirements/** - Requirements, user stories, acceptance criteria
+2. **2-analysis/** - Risk assessment, feasibility studies, impact analysis  
+3. **3-architecture/** - Design overview, technical design, integration points
+4. **4-development/** - Implementation logs, code decisions, testing notes
+5. **5-debugging/** - Issues logs, solutions, troubleshooting guides
+6. **6-key_learnings/** - Lessons learned, best practices, future improvements
+
+## Related Documentation
+
+- **Project Config**: `../../.brtops/feature-registry.json`
+- **After Action Reports**: `../04_after-action-reports/`
+- **Quality Standards**: `../05_quality/`
+- **Agent Context**: `../06_for-agents/`
+
+---
+
+**Last Updated**: 2025-08-21  
+**BRTOPS Version**: 1.1.002  
+**Registry Status**: Complete