--- a/docs/04_after-action-reports/app-header-button-tweaks_2025-08-21.md
+++ b/docs/04_after-action-reports/app-header-button-tweaks_2025-08-21.md
@@ -1,0 +1,28 @@
+---- AFTER ACTION REPORT | App Header Button Tweaks ----
+DATE: 2025-08-21
+
+CONTRIBUTORS: Branden Thompson & Claude (AI Agent)
+
+FINAL RESULT: SUCCESS 
+
+TL;DR: Successfully converted all Active Conversation Controls in app header to icon-only variants with accessibility tooltips, maintaining functionality while improving UI cleanliness.
+
+ACTIONS:
+ ✅ SUCCESS: Play/Resume button converted to icon-only with tooltip
+ ✅ SUCCESS: Pause button converted to icon-only with tooltip  
+ ✅ SUCCESS: Stop button converted to icon-only with tooltip
+ ✅ SUCCESS: Scope validation - Card Controls correctly excluded from scope
+ ✅ SUCCESS: Quality gates validation (build, lint, functionality)
+ ✅ SUCCESS: Git commit with BRTOPS formatting
+ ✅ SUCCESS: Automated dev page updates via post-commit hooks
+ ✅ SUCCESS: Clean dev testing using ./dev-scripts/setup/clean-start-dev.sh
+
+LEARNINGS: 
+BRTOPS v1.1.001 directive commands proved highly effective for scope management, preventing feature creep through clear boundary definition. The DIRECTIVE command successfully established persistent operational instructions equivalent to Claude '#' commands. Project's automated quality integration via post-commit hooks maintained dev page accuracy without manual intervention. Clean development startup script ensured single, isolated testing environment eliminating port conflicts and cross-tab synchronization issues.
+
+DOCUMENTATION:
+ 1. /Users/bthompso/Desktop/PERSONAL_PROJECTS/conversation-tracker/components/ui/conversation-controls.jsx
+ 2. /Users/bthompso/Desktop/PERSONAL_PROJECTS/conversation-tracker/docs/04_after-action-reports/app-header-button-tweaks_2025-08-21.md
+
+NOTES AND ACTION ITEMS:
+Feature successfully demonstrates icon-only conversion pattern for future UI tweaks. BRTOPS v1.1.001 integration during live development validated new directive system effectiveness.