--- a/docs/04_after-action-reports/documentation-migration-brtops-v1102_2025-08-21.md
+++ b/docs/04_after-action-reports/documentation-migration-brtops-v1102_2025-08-21.md
@@ -1,0 +1,33 @@
+---- AFTER ACTION REPORT | Documentation Migration BRTOPS v1.1.002 ----
+DATE: 2025-08-21
+
+CONTRIBUTORS: Branden Thompson & Claude (AI Agent)
+
+FINAL RESULT: SUCCESS 
+
+TL;DR: Successfully migrated conversation-tracker project documentation to full BRTOPS v1.1.002 compliance, establishing universal project structure standards and improving organizational consistency.
+
+ACTIONS:
+ ✅ SUCCESS: Audited current documentation structure against BRTOPS v1.1.002 standards
+ ✅ SUCCESS: Designed comprehensive migration plan with folder mapping strategy
+ ✅ SUCCESS: Renamed docs/03_hygiene → docs/03_operations per BRTOPS standards
+ ✅ SUCCESS: Created docs/05_quality/ with testing, quality-gates, reviews folders
+ ✅ SUCCESS: Renumbered for-agents (05→06) and archive (06→07) folders
+ ✅ SUCCESS: Updated docs/README.md with complete BRTOPS v1.1.002 structure
+ ✅ SUCCESS: Created testing strategy documentation and quality procedures
+ ✅ SUCCESS: Validated 100% BRTOPS compliance across all folder structures
+ ✅ SUCCESS: Preserved all existing documentation content during migration
+
+LEARNINGS: 
+The conversation-tracker project already had strong documentation organization that aligned closely with BRTOPS principles, requiring minimal structural changes. The primary gaps were missing quality documentation folder and inconsistent naming conventions. BRTOPS v1.1.002 universal structure provides excellent framework for maintaining documentation consistency across projects of different scales and types.
+
+The migration demonstrated that existing well-organized projects can be adapted to BRTOPS standards without disrupting established workflows. The addition of formal quality documentation structure enhances project governance while the standardized folder hierarchy improves navigation and discoverability for both human team members and AI agents.
+
+DOCUMENTATION:
+ 1. /Users/bthompso/Desktop/PERSONAL_PROJECTS/conversation-tracker/docs/README.md
+ 2. /Users/bthompso/Desktop/PERSONAL_PROJECTS/conversation-tracker/docs/05_quality/README.md
+ 3. /Users/bthompso/Desktop/PERSONAL_PROJECTS/conversation-tracker/docs/05_quality/testing/testing-strategy.md
+ 4. /Users/bthompso/Desktop/PERSONAL_PROJECTS/conversation-tracker/docs/04_after-action-reports/documentation-migration-brtops-v1102_2025-08-21.md
+
+NOTES AND ACTION ITEMS:
+Migration establishes foundation for consistent BRTOPS adoption across all future projects. Quality documentation structure now available for enhanced testing and review procedures. Universal structure compliance enables better AI agent navigation and understanding of project organization.