… docs/03_operations/2025-08-18-bag-n-tag-cleanup.md
… docs/03_operations/2025-08-18-build-cleanup.md
… docs/03_operations/2025-08-18-data-flow-architecture-analysis.md
… docs/03_operations/2025-08-18-documentation-organization.md
… docs/03_operations/2025-08-18-main-app-navigation-rename.md
… docs/03_operations/2025-08-18-navigation-label-clarification.md
… docs/03_operations/2025-08-18-react-query-migration-completion.md
… docs/03_operations/2025-08-18-react-query-week1-implementation.md
… docs/03_operations/2025-08-18-session-card-styling.md
… docs/03_operations/2025-08-18-synthwave84-theme-implementation.md
… docs/03_operations/2025-08-18-ui-constants-card-background.md
… docs/03_operations/2025-08-18-user-switching-bug-fix.md
… docs/03_operations/2025-08-19-collaboration-api-runaway-learnings.md
… docs/03_operations/2025-08-19-ui-theming-bag-n-tag.md
… docs/03_operations/2025-08-20-active-stackers-sse-migration-checkpoint.md
… docs/03_operations/2025-08-20-active-stackers-sse-migration-success.md
… docs/03_operations/2025-08-20-active-stackers-sync-fix.md
… docs/03_operations/2025-08-20-dev-pages-theme-awareness.md
… docs/03_operations/2025-08-20-hardcoded-color-audit-final-report.md
… docs/03_operations/2025-08-20-sev-0-sse-emergency-fix.md
… docs/03_operations/2025-08-20-sse-emergency-fix-FAIL-STATE.md
… docs/03_operations/2025-08-20-theme-architecture-verification.md
… docs/03_operations/2025-08-21-active-stackers-flickering-comprehensive-resolution.md
… docs/03_operations/2025-08-21-active-stackers-flickering-debug-analysis.md
… docs/03_operations/2025-08-21-active-stackers-verification-testing-success.md
… docs/03_operations/2025-08-21-sse-infrastructure-optimization.md
… docs/03_operations/2025-08-21-sse-multi-tab-debugging-analysis.md
… docs/03_operations/2025-08-21-sse-multi-tab-pdsop-bag-n-tag-assessment.md
… docs/03_operations/2025-08-21-sse-optimization-testing-ready.md
… docs/03_operations/README.md